master-server/game-box-master-server-unity/src/main/java/com/assambra/gameboxmasterserverunity/manager/SimpleRoomManager.java
package com.assambra.gameboxmasterserverunity.manager;

import com.assambra.gameboxmasterserverunity.entity.Room;

public class SimpleRoomManager<R extends Room> extends AbstractRoomManager<R> {

    public SimpleRoomManager() {
        this(10000);
    }

    public SimpleRoomManager(int maxRoom) {
        super(maxRoom);
    }

    protected SimpleRoomManager(Builder<?, ?> builder) {
        super(builder);
    }

    public static Builder<?, ?> builder() {
        return new Builder<>();
    }

    public static class Builder<R extends Room, B extends Builder<R, B>>
        extends AbstractRoomManager.Builder<R, B> {

        @Override
        public SimpleRoomManager<R> build() {
            return new SimpleRoomManager<>(this);
        }
    }
}
